refactor(Move Calcualte Manager): 계산기 매니저를 로또 기계 안에 위치
- 결국 해당 계산기도 로또 기계 안에서 실행되는 것이라 판단하여
- 로또 기계 안에 위치
